sql: split API for ephemeral spaces creation
tarantoolSqlite3EphemeralCreate was creating ephemeral space
and was setting provided cursor to the first entry. In some cases
when we insert data into ephemeral space, cursor is unnecessary,
the only necessary thing is a pointer to struct space. In this
patch tarantoolSqlite3EphemeralCreate was divided into two
methods: tarantoolSqlite3EphemeralCreate, which only creates a
ephemeral space and tarantoolSqlite3EphemeralSetFirst, which sets
cursor to the first entry.

For #3116