sql: change OP_IdxInsert semantics
For performing write operations like insertion, a whole write
cursor is completely unnecessary, the only necessary information
for Tarantool are pointers to space, tuple and end of tuple. The
whole BtCursor structure is useless. However, if you write some
data into ephemeral space, sooner or later data will be extracted
from it, therefore read cursor is useful in that case.

In this patch IdxInsert opcode semantics was changed. Flag in P5
register now determines in what kind of space the insertion is
performed and how IdxInsert treat number in P1 - as a cursor
number (if space is ephemeral) and as a memory cell number if
space is ordinary. In addition Tarantool SQL API for insertion
was refactored according to the changes described above.

For #3116