sql: add index_def to Index, fixed Index->prop to Index->def->prop
Now every sqlite struct Index is created with tnt struct
index_def inside. This allows us to use tnt index_def in work
with sqlite indexes in the same manner as with tnt index and is a
step to remove sqlite Index with tnt index.
However sqlite Index and tnt index are not interchangeable
(e.g. sqlite index use -1 for index part that represent
pk integer column but tnt index part can’t do that as
they are unsigned ints).

Needed for: #3369