vinyl: relax limitation imposed on run min/max lsn
Currently, we assume that no two runs of the same range intersect by
LSN. This holds, because LSNs grow strictly monotonically, and no
transaction may be split between two runs (as we pin each affected
vy_mem until the transaction is complete). We ensure this with an
assertion in vy_task_dump_complete.

However, when building a new index we can't increment tx_manager->lsn so
there may be multiple statements with the same LSN. This is OK as for
each particular key, two statements will still have different LSNs, but
this may break the assertion in vy_task_dump_complete in case dump
occurs while build is in progress.

To avoid that, let's relax the condition under the assertion and assume
that a run's max_lsn may be equal min_lsn of the newer run.

Needed for #1653